Read Atoms.toml, generate rust, build
Looks good as a prototype. Failing, however, because CustomAtoms (the list object for holding all of the things) is specifying a Box<UcumSymbol> to all its Atom elements to be generic, but I'm getting the "the trait `ucum_symbol::UcumSymbol` cannot be made into an object" error. Could probably fix, but it's made me reconsider using a trait object here.