AGDEV-32978 Remove Unit::div_32 and Unit::mul_32
It doesn't really make sense to have these (what purpose do they serve?). This lent itself to problems too: if you have a unit that is `10m` and you divide that by 3, since UCUM says the factor (the `10` in the original unit) must be an integer, you end up rounding that to `3m`, which just seems wrong.