AGDEV-32978 Div/Mul refactor; Add Convertible trait and implement
First, I found out that it's possible to `impl Div<f64>`, which let me do away with `Measurement::div_scalar()` (etc.). Next I added the `Convertible` trait, which let me define `convert_to` for both a str param and a Unit param. Doing that let me refactor some places where we were forcing a parse of a string just to create a `Unit`, when we already had the `Unit` we needed.