DEV-331 Redesign Composition; fix bug for ticket
Did the redesign of the `Composition` struct because the old way (using an internal `BTreeMap` for building the `Composition`) a) was a bit difficult to manage the `Entry` API in this case, b) was less precise than it could be--a `Composition` can _only_ ever have keys that match `Dimension`s so now it has just that.

Also redid how `Composition`s are combined with each other. Previously this was done from the `Term` perspective; now logic is encapsulated in `Composition` via its implementation of `Mul` and `Mul<i32>`. This made the bug easier to track down.