Merged in feature/remove_reprc_structs (pull request #52)
Remove Measurement and Unit wrapper structs.

cbindgen produces a struct definition for the wrapper types, but it
includes the private field definitions whose types *don't* have a
definition in the header.

Clients don't interact directly with the structs (they just pass the
pointer back in to wise_units-ffi functions), so we don't need a type
that's laid out in memory for C interop; just treat the `Box`ed struct
as an `OpaquePointer` on the client side.

Approved-by: Randy Stoller <randy@agrian.com>
Approved-by: Steve Loveless <steve@agrian.com>