Add support for deriving `Measurement` and `Unit` FFIs.
`Measurement`'s FFI is all derived automatically, but `Unit` is a bit
different; we don't care to expose its `terms` (never mind the fields of
`Term`) over FFI, so we need to be explicit about how this type is
initialized and how its fields are accessed over FFI. (The `ffi_consumer`
side of this is pretty hacky for now; it pretty much only supports this
scenario until we refactor it in DEV-13175).

Note that for now we also have two FFIs for wise_units; the manually
written one in `wise_units-ffi`, and the `ffi_derive` one in
`wise_units` (when the `cffi` feature is on). We should be able to replace
`wise_units-ffi` completely at some point, but for now the derived FFI is
just here to support deriving an FFI for `agrian_types::CommodityLocation`.

Add .cargo/config.toml so that this crate knows where to find its
dependencies without needing a particular environment configuration.