Add benchmarks; AST transformer
I think the parser is fast and good, but it's turning the AST into our current set of Terms and such that takes a while (probably the matching on an atom/prefix's primary code to look up our current Atom variant?). Punting this until a better time to make bigger changes.